Create CIS129_AmberUidenich_Lab6.py
This is the work for Module 6 Lab.
The code calculates the number of hotdogs needed for a cookout based on number of people and a hot dogs per person count. It then takes that number and calculates the minimum number of packages of hotdogs and buns need to be purchased. Finally, it calculates how many hotdogs and buns will be left over.